refactor(xref-repair): own module, tokenized trailer walk, EOF-anchored startxref
Review follow-ups on the short-entry repair:

- Move the repair and its lexical helpers out of lib.rs into
  src/xref_repair.rs.
- find_dict_key walks the trailer as key/value pairs and skips each
  value whole (nested dictionaries and arrays, literal and hex strings,
  comments, `n g R` references), so `/Prev` or `/XRefStm` inside a value
  is never taken for the key.
- pdf_dictionary_span skips `%` comments and hex strings, so a `>>`
  inside either no longer ends the dictionary early.
- last_startxref_offset anchors on the document's final `%%EOF`
  wherever it is, requires the keyword to start a line (leading spaces
  or tabs allowed), skips candidates whose offset is missing, unparsable
  or out of range instead of giving up, and does it in one backwards
  walk.
